requirement.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.2
pycparser==2.20
PyMySQL==1.0.2
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.1
wtforms-validators==1.0.0
